Add files via upload
A program that compares two text files whose names will be passed to it as arguments to the call. This comparison should be carried out line by line, listing on the screen those lines that differ in content, along with their number and the name of the file from which they originate. The number should be assigned to the line relative to the beginning of the file, i.e. the first line should be number one, the second line two, etc.